Add files via upload
Data from normalize SQL tables are imported into Microsoft Power BI to be used to create key performance indicators and metrics for a medical laboratory. Microsoft Power BI is used to transform and clean the data using the Microsoft M and DAX languages and subsequent merged tables are created to create eloquent laboratory enterprise dashboards to create business insights as seen in this file.